refactor(neighbor): relocate cognition modules under familiar_neighbor.mind
Move 14 cognition modules from src/familiar_agent/ to
src/familiar_neighbor/mind/, replacing the old paths with explicit
compatibility shims so existing imports keep working:

  appraisal, attention_schema, concern_engine, default_mode, desires,
  interoception, meta_monitor, prediction, relationship, scene,
  self_narrative, self_state, social_policy, workspace

The shims re-export every public symbol via ``from familiar_neighbor.mind.X
import *``; the original test surface in tests/test_*.py keeps working
unchanged.  ``mental_state.py`` stays in ``familiar_agent`` for now: it is
shared with non-neighbour callers (interoception / appraisal pull
``InteroceptiveSignal`` and ``AffectiveState`` from it), and a
coordinated move can land alongside the agent.py hook rewrite.

Internal imports that previously used the ``.<sibling>`` form for
``mental_state`` / ``_i18n`` / ``sqlite_migrations`` (modules that
stayed in familiar_agent) are rewritten to absolute
``familiar_agent.<sibling>`` imports.

Updated tests:

- tests/test_desire_v2.py: 13 ``patch("familiar_agent.desires.datetime")``
  invocations rewritten to patch the new location
  (``familiar_neighbor.mind.desires.datetime``) because the shim does not
  forward attribute writes back to the implementation module.

Validation: ruff/format/mypy/pytest all clean (951 passed).
